v0.6.5 - fixed reflection bug.
There was a subtle bug with the reflection from the walls and the paddles.
It was not manifested often. The reason for it was that I forgot to move
the ball a bit backwards after the collision has occurred, and this
sometimes lead to instabilities and the ball flew away from the box.